restructure tests
Changed most file tests to be golden tests, avoids reinventing the wheel and
means that any new desired golden tests are trivial to add.

Moved bad input files to their own folder.

Renamed the unremarkable 'testN' files to something a bit more descriptive.

moved golden files to be subfolder of existing 'json-data' folder to avoid
things spreading out in a weird way. The json files should now all be located in
a single folder.

update the respective cabal files to include the new 'extra-source-file'
information.
