+++ test/json-data/goldens/empty_arr_empty_ws.json.golden
