+++ test/json-data/goldens/nested_arrs.json.golden
