+++ test/json-data/goldens/null_arr_trailing_comma_ws.json.golden
