+++ test/json-data/goldens/unicode_2705.json.golden
@@ -1,0 +1,1 @@
+"\u2705"
