+++ waargbench/waargbench.cabal
@@ -35,15 +35,9 @@ tested-with:         GHC==7.10.3
 
 -- Extra files to be distributed with the package, such as examples or a
 -- README.
-extra-source-files:  test/json-data/jp100.json
+extra-source-files:  test/json-data/twitter_with_hex_vals.json
                    , test/json-data/numbers.json
-                   , test/json-data/test1.json
-                   , test/json-data/test2.json
-                   , test/json-data/test3.json
-                   , test/json-data/test4.json
-                   , test/json-data/test5.json
-                   , test/json-data/test6.json
-                   , test/json-data/test7.json
+                   , test/json-data/image_obj.json
                    , test/json-data/twitter100.json
 
 benchmark bench
